feat: Add Home and ErrorNotFoundPage components
This commit adds the Home and ErrorNotFoundPage components to the frontend application. The Home component renders a simple "Home" heading, while the ErrorNotFoundPage component displays a "404" heading and a "Page not found" message. These components enhance the user interface of the application and improve the overall user experience.
