+++ frontend/src/pages/error/notFoundPage/ErrorNotFoundPage.tsx
@@ -1,0 +1,10 @@
+const ErrorNotFoundPage = () => {
+    return (
+        <div>
+            <h1>404</h1>
+            <p>Page not found</p>
+        </div>
+    );
+}
+
+export default ErrorNotFoundPage;
