+++ frontend/src/pages/home/Home.tsx
@@ -1,0 +1,9 @@
+const Home = () => {
+    return (
+        <div>
+            <h1>Home</h1>
+        </div>
+    )
+}
+
+export default Home;
